requirements.txt
asgiref==3.8.1
Django==5.1
django-auto-logout==0.5.1
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==11.1.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.1
